backend/admin/requirements.txt
psycopg2-binary==2.9.9
pyjwt==2.8.0
pydantic==2.5.0
